+++ api/src/main/java/com/marketplace/dto/category/CategoryResponseDto.java
@@ -1,16 +1,3 @@
 package com.marketplace.dto.category;
 
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor
-public class CategoryResponseDto {
-
-    private Long id;
-    private String name;
-}
+public record CategoryResponseDto (Long id, String name) { }
