+++ api/src/main/java/com/marketplace/dto/category/CategoryUpdateDto.java
@@ -1,17 +1,10 @@
 package com.marketplace.dto.category;
 
 import io.swagger.v3.oas.annotations.media.Schema;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
+import jakarta.validation.constraints.NotBlank;
 
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor
-public class CategoryUpdateDto {
-
+public record CategoryUpdateDto (
     @Schema(description = "Nom de la catégorie", example="fruits et légumes")
-    private String name;
-}
+    @NotBlank(message = "Le nom de la catégorie ne peut pas être vide")
+    String name
+) {}
